requirements.txt
web3==5.12.1
click==7.1.2
pyyaml==5.3.1
